storage: added fixtures sub-package
