+++ services/storage/tests/conftest.py
@@ -1,9 +1,9 @@
-# pylint:disable=unused-variable
-# pylint:disable=unused-argument
-# pylint:disable=redefined-outer-name
-# pylint:disable=unsupported-assignment-operation
-# pylint:disable=no-name-in-module
-# pylint:disable=no-member
+# pylint: disable=no-member
+# pylint: disable=no-name-in-module
+# pylint: disable=redefined-outer-name
+# pylint: disable=unsupported-assignment-operation
+# pylint: disable=unused-argument
+# pylint: disable=unused-variable
 
 
 import datetime
@@ -18,7 +18,7 @@ from typing import Dict, Iterator, Tuple
 import dotenv
 import pytest
 import simcore_service_storage
-import utils
+import tests.utils
 from aiohttp import web
 from aiopg.sa import create_engine
 from servicelib.application import create_safe_application
@@ -26,7 +26,7 @@ from simcore_service_storage.datcore_wrapper import DatcoreWrapper
 from simcore_service_storage.dsm import DataStorageManager, DatCoreApiToken
 from simcore_service_storage.models import FileMetaData
 from simcore_service_storage.settings import SIMCORE_S3_STR
-from utils import (
+from tests.utils import (
     ACCESS_KEY,
     BUCKET_NAME,
     DATA_DIR,
@@ -36,6 +36,10 @@ from utils import (
     USER,
     USER_ID,
 )
+
+pytest_plugins = [
+    "tests.fixtures.data_models",
+]
 
 CURRENT_DIR = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
 
@@ -130,7 +134,7 @@ def postgres_service(docker_services, docker_ip):
 
     # Wait until service is responsive.
     docker_services.wait_until_responsive(
-        check=lambda: utils.is_postgres_responsive(url),
+        check=lambda: tests.utils.is_postgres_responsive(url),
         timeout=30.0,
         pause=0.1,
     )
@@ -159,11 +163,11 @@ def postgres_service_url(postgres_service, docker_services, docker_ip):
         port=docker_services.port_for("postgres", 5432),
     )
 
-    utils.create_tables(url)
+    tests.utils.create_tables(url)
 
     yield url
 
-    utils.drop_tables(url)
+    tests.utils.drop_tables(url)
 
 
 @pytest.fixture(scope="function")
@@ -191,7 +195,7 @@ def minio_service(docker_services, docker_ip):
 
     # Wait until service is responsive.
     docker_services.wait_until_responsive(
-        check=lambda: utils.is_responsive(url, 403),
+        check=lambda: tests.utils.is_responsive(url, 403),
         timeout=30.0,
         pause=0.1,
     )
@@ -243,7 +247,7 @@ def mock_files_factory(tmpdir_factory):
 @pytest.fixture(scope="function")
 def dsm_mockup_complete_db(postgres_service_url, s3_client) -> Tuple[Dict, Dict]:
 
-    utils.fill_tables_from_csv_files(url=postgres_service_url)
+    tests.utils.fill_tables_from_csv_files(url=postgres_service_url)
 
     bucket_name = BUCKET_NAME
     s3_client.create_bucket(bucket_name, delete_contents_if_exists=True)
@@ -341,7 +345,7 @@ def dsm_mockup_db(
         data[object_name] = FileMetaData(**d)
 
         # pylint: disable=no-member
-        utils.insert_metadata(postgres_service_url, data[object_name])
+        tests.utils.insert_metadata(postgres_service_url, data[object_name])
 
     total_count = 0
     for _obj in s3_client.list_objects_v2(bucket_name, recursive=True):
